% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationCalculationFunctions.R
\name{calcCorrelationLengths}
\alias{calcCorrelationLengths}
\title{Get correlation lengths from profiles}
\usage{
calcCorrelationLengths(groupedProfiles)
}
\arguments{
\item{correlationProfiles}{}
}
\description{
This function takes a data frame (with grouping) and determines the zero-intercept (correlation length) for each profile.
It is dependent on nesting ahd purrr map functions to iterate over all the profiles.
}
