% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalOrderMeasureFunctions.R
\name{measureDilatation}
\alias{measureDilatation}
\title{Collective dilatation order}
\usage{
measureDilatation(frameData)
}
\arguments{
\item{frameData}{A data frame with the columns R, C, u, and V specifying optical flow properties}
}
\value{
A scalar quantifying the dilatation order (as in Couzin et al. 2002)
}
\description{
This is based off of Attanasi et al. 2014 paper on swarmes of midges where they define an order parameter that measures the extent to which
particles are moving outward (positive) or inward (negative) toward the center of mass of the collective.
}
