% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalOrderMeasureFunctions.R
\name{measureInternalOrder}
\alias{measureInternalOrder}
\title{Title}
\usage{
measureInternalOrder(dataFolder, frameChunkSize = 3600, dplyr = FALSE)
}
\arguments{
\item{dplyr}{Is the dplyr package available? This shouldn't be provided by the user but should be checked.}
}
\description{
Title
}
