% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalOrderMeasureFunctions.R
\name{measurePolarization}
\alias{measurePolarization}
\title{Collective polarization order}
\usage{
measurePolarization(frameData)
}
\arguments{
\item{frameData}{A data frame with the columns R, C, u, and V specifying optical flow properties}
}
\value{
A scalar quantifying the polarization order (as in Couzin et al. 2002)
}
\description{
Collective polarization order
}
